+++ samples/GenericInterface.ts
@@ -1,0 +1,13 @@
+export interface Generic {
+
+    longs: Array<number>;
+
+    doubles: Array<number>;
+
+    strings: Array<string>;
+
+    floats: Array<number>;
+
+    booleans: Array<boolean>;
+
+}