+++ samples/OtherInterface.ts
@@ -1,0 +1,7 @@
+import {Simple} from './Simple';
+
+export interface Other {
+
+    simple: Simple;
+
+}