+++ samples/SimpleInterface.ts
@@ -1,0 +1,7 @@
+export interface Simple {
+
+    number: number;
+
+    string: string;
+
+}