+++ samples/hierarchy/FemaleInterface.ts
@@ -1,0 +1,7 @@
+import {Human} from './Human';
+
+export interface Female extends Human {
+
+    type: string;
+
+}