+++ samples/hierarchy/MaleInterface.ts
@@ -1,0 +1,7 @@
+import {Human} from './Human';
+
+export interface Male extends Human {
+
+    type: string;
+
+}