+++ src/ast/Field.js
@@ -1,10 +1,10 @@
 const Type = require('./Type');
 
-module.exports.generate = (from) => {
+module.exports.generate = (from, generateHasClass) => {
     let {name, scope, type} = from;
     const t = Type(type);
     return {
-        str: `${scope} ${name}: ${t.name};`,
+        str: `${generateHasClass ? `${scope} ` : ''}${name}: ${t.name};`,
         needImport: t.needImport
     };
 };