+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * `1.2.0` :
   * add newline to generated files
   * add enum
+  * generate has interface (default)
 * `1.1.0` :
   * add generics
   * add type :